// SwarmPortal.FrontEnd/src/app/api/model/iGroup.ts
/**
 * SwarmPortal.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Link } from './link';


export interface IGroup { 
    id?: number;
    name?: string | null;
    links?: Array<Link> | null;
}